CAMEL-20113: camel/tests consolidating org.apache.camel.itest.jaxb: -SpringLoadRouteFromXmlTest - moved to the camel-jaxb component; -JaxbFallbackTypeConverterTest - moved to the camel-jaxb component;
org.apache.camel.itest.jetty:
-JettyBridgeHostHeaderIssueTest - moved to camel-jetty;
-JettyConstantSetHeaderTest - setting headers is checked in other camel-jetty tests, so it is deleted;
-JettyFailoverRoundRobinTest - moved to camel-jetty;
-JettyJmsShutdownInProgressTest - jetty graceful shutdown is checked in org.apache.camel.component.jetty.JettySuspendWhileInProgressTest, so it is deleted;
-JettyJmsTest - test was made just to show example route, so it is deleted;
-JettyJmsTwowayTest - test was made just to show example route, so it is deleted;
-JettyRestRedirectTest - moved to camel-jetty;
-JettySimpleSetHeaderTest - setting headers is checked in other camel-jetty tests, so it is deleted;
-JettyValidatorStreamTest - moved to camel-jetty;
-JettyValidatorTest - moved to camel-jetty;
-JettyXsltHttpTemplateTest - moved to camel-jetty;
-JettyXsltTest - moved to camel-jetty;
