+++ components/camel-jaxb/src/test/resources/org/apache/camel/example/BarRoute.xml
