+++ components/camel-jetty/src/test/java/org/apache/camel/component/jetty/JettyValidatorTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.itest.jetty;
+package org.apache.camel.component.jetty;
 
 import java.io.InputStream;
