+++ components/camel-jetty/src/test/java/org/apache/camel/component/jetty/rest/RestJettyRedirectTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.itest.jetty;
+package org.apache.camel.component.jetty.rest;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class JettyRestRedirectTest extends CamelTestSupport {
+public class RestJettyRedirectTest extends CamelTestSupport {
 
     private int port;
