+++ components/camel-jetty/src/test/resources/OptimizationRequest.xsd
