+++ components/camel-jetty/src/test/resources/org/apache/camel/component/jetty/JettySpringFailoverRoundRobinTest.xml
